Added function map-query-alignments in src/name.lisp
The new function `map-query-alignments' is intended as the foundation
for all functions that match queries against (sets of) names, where
both the query and the match candidates may contain wild
components.

In particular, the complexity of the to-be-added function
`map-matching-options' can be kept under control by using
`map-options' in combination with `map-query-alignments'.

* src/name.lisp (map-query-alignments): new function; call a function
  with all alignments of two potentially wild names
* test/name.lisp (test map-query-alignments.smoke): new test; smoke
  test for the `map-query-alignments' function
* test/package.lisp (package configuration.options.test): added
  import-from configuration.options symbol map-query-alignments